Improve JPA Embeddable FAT to run on OpenLiberty only
