+++ dev/com.ibm.ws.jpa.tests.spec10.embeddable_fat/fat/src/com/ibm/ws/jpa/spec10/embeddable/FATSuite.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2021 IBM Corporation and others.
+ * Copyright (c) 2019, 2022 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -11,45 +11,17 @@
 
 package com.ibm.ws.jpa.spec10.embeddable;
 
-import org.junit.ClassRule;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
-import org.testcontainers.containers.JdbcDatabaseContainer;
 
-import com.ibm.ws.jpa.spec10.embeddable.issues.TestOLGH10310_EJB;
-import com.ibm.ws.jpa.spec10.embeddable.issues.TestOLGH10310_Web;
-
-import componenttest.containers.ExternalTestServiceDockerClientStrategy;
-import componenttest.rules.repeater.RepeatTests;
-import componenttest.topology.database.container.DatabaseContainerFactory;
+import com.ibm.ws.jpa.spec10.embeddable.tests.AbstractFATSuite;
 
 @RunWith(Suite.class)
 @SuiteClasses({
-                TestOLGH10310_Web.class,
-                TestOLGH10310_EJB.class,
-                JPA10EmbeddableBasic_EJB.class,
-                JPA10EmbeddableBasic_WEB.class,
-                JPA10EmbeddableNested_EJB.class,
-                JPA10EmbeddableNested_WEB.class,
-                JPA10EmbeddableRelationship_EJB.class,
-                JPA10EmbeddableRelationship_WEB.class,
+                JPA20Suite.class,
                 componenttest.custom.junit.runner.AlwaysPassesTest.class
 })
-public class FATSuite {
-    public final static String[] JAXB_PERMS = { "permission java.lang.RuntimePermission \"accessClassInPackage.com.sun.xml.internal.bind.v2.runtime.reflect\";",
-                                                "permission java.lang.RuntimePermission \"accessClassInPackage.com.sun.xml.internal.bind\";",
-                                                "permission java.lang.RuntimePermission \"accessDeclaredMembers\";" };
+public class FATSuite extends AbstractFATSuite {
 
-    //Required to ensure we calculate the correct strategy each run even when
-    //switching between local and remote docker hosts.
-    static {
-        ExternalTestServiceDockerClientStrategy.setupTestcontainers();
-    }
-
-    @ClassRule
-    public static final JdbcDatabaseContainer<?> testContainer = DatabaseContainerFactory.create();
-
-    @ClassRule
-    public static RepeatTests r = RepeatTests.with(new RepeatWithJPA20());
 }
