+++ dev/com.ibm.ws.jpa.tests.spec10.embeddable_fat/fat/src/com/ibm/ws/jpa/spec10/embeddable/issues/TestOLGH10310_Web.java
@@ -29,7 +29,7 @@ import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
 import com.ibm.ws.jpa.olgh10310.web.TestOLGH10310Servlet;
 import com.ibm.ws.jpa.spec10.embeddable.FATSuite;
-import com.ibm.ws.jpa.spec10.embeddable.JPAFATServletClient;
+import com.ibm.ws.jpa.spec10.embeddable.tests.JPAFATServletClient;
 
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
