+++ dev/com.ibm.ws.jpa.tests.spec10.embeddable_fat/fat/src/com/ibm/ws/jpa/spec10/embeddable/tests/JPA10EmbeddableBasic_WEB.java
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.spec10.embeddable;
+package com.ibm.ws.jpa.spec10.embeddable.tests;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -26,10 +26,11 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
+import com.ibm.websphere.simplicity.config.ClassloaderElement;
+import com.ibm.websphere.simplicity.config.ConfigElementList;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.embeddable.nested.ejb.TestEmbeddableNested_EJB_SFEx_Servlet;
-import com.ibm.ws.jpa.embeddable.nested.ejb.TestEmbeddableNested_EJB_SF_Servlet;
-import com.ibm.ws.jpa.embeddable.nested.ejb.TestEmbeddableNested_EJB_SL_Servlet;
+import com.ibm.ws.jpa.embeddable.basic.web.TestEmbeddableBasicServlet;
+import com.ibm.ws.jpa.spec10.embeddable.FATSuite;
 
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
@@ -44,11 +45,12 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class JPA10EmbeddableNested_EJB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "embeddableNestedEjb";
-    private final static String RESOURCE_ROOT = "test-applications/embeddable/nested/";
-    private final static String appFolder = "ejb";
-    private final static String appName = "embeddableNestedEjb";
+public class JPA10EmbeddableBasic_WEB extends JPAFATServletClient {
+
+    private final static String CONTEXT_ROOT = "embeddableBasicWeb";
+    private final static String RESOURCE_ROOT = "test-applications/embeddable/basic/";
+    private final static String appFolder = "web";
+    private final static String appName = "embeddableBasicWeb";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
@@ -57,15 +59,13 @@ public class JPA10EmbeddableNested_EJB extends JPAFATServletClient {
     private static long timestart = 0;
 
     static {
-        dropSet.add("JPA10_EMBEDDABLE_NESTED_DROP_${dbvendor}.ddl");
-        createSet.add("JPA10_EMBEDDABLE_NESTED_CREATE_${dbvendor}.ddl");
+        dropSet.add("JPA10_EMBEDDABLE_BASIC_DROP_${dbvendor}.ddl");
+        createSet.add("JPA10_EMBEDDABLE_BASIC_CREATE_${dbvendor}.ddl");
     }
 
     @Server("JPA10Server")
     @TestServlets({
-                    @TestServlet(servlet = TestEmbeddableNested_EJB_SL_Servlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableNested_EJB_SL_Servlet"),
-                    @TestServlet(servlet = TestEmbeddableNested_EJB_SF_Servlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableNested_EJB_SF_Servlet"),
-                    @TestServlet(servlet = TestEmbeddableNested_EJB_SFEx_Servlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableNested_EJB_SFEx_Servlet")
+                    @TestServlet(servlet = TestEmbeddableBasicServlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableBasicServlet")
     })
     public static LibertyServer server;
 
@@ -74,7 +74,7 @@ public class JPA10EmbeddableNested_EJB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(JPA10EmbeddableNested_EJB.class);
+        bannerStart(JPA10EmbeddableBasic_WEB.class);
         timestart = System.currentTimeMillis();
 
         int appStartTimeout = server.getAppStartTimeout();
@@ -99,7 +99,7 @@ public class JPA10EmbeddableNested_EJB extends JPAFATServletClient {
 
         final Set<String> ddlSet = new HashSet<String>();
 
-        System.out.println(JPA10EmbeddableNested_EJB.class.getName() + " Setting up database tables...");
+        System.out.println(JPA10EmbeddableBasic_WEB.class.getName() + " Setting up database tables...");
 
         ddlSet.clear();
         for (String ddlName : dropSet) {
@@ -117,20 +117,15 @@ public class JPA10EmbeddableNested_EJB extends JPAFATServletClient {
     }
 
     private static void setupTestApplication() throws Exception {
-        JavaArchive ejbApp = ShrinkWrap.create(JavaArchive.class, appName + ".jar");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.embeddable.nested.ejblocal");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.embeddable.nested.model");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.embeddable.nested.testlogic");
-        ShrinkHelper.addDirectory(ejbApp, RESOURCE_ROOT + appFolder + "/" + appName + ".jar");
-
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.nested.ejb");
+        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.basic.model");
+        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.basic.testlogic");
+        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.basic.web");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
 
         final EnterpriseArchive app = ShrinkWrap.create(EnterpriseArchive.class, appNameEar);
-        app.addAsModule(ejbApp);
         app.addAsModule(webApp);
         app.addAsLibrary(testApiJar);
         ShrinkHelper.addDirectory(app, RESOURCE_ROOT + appFolder, new org.jboss.shrinkwrap.api.Filter<ArchivePath>() {
@@ -150,11 +145,12 @@ public class JPA10EmbeddableNested_EJB extends JPAFATServletClient {
         Application appRecord = new Application();
         appRecord.setLocation(appNameEar);
         appRecord.setName(appName);
-//        ConfigElementList<ClassloaderElement> cel = appRecord.getClassloaders();
-//        ClassloaderElement loader = new ClassloaderElement();
-//        loader.setApiTypeVisibility("+third-party");
-////        loader.getCommonLibraryRefs().add("HibernateLib");
-//        cel.add(loader);
+
+        // For OpenJPA, Oracle CLOB support
+        ConfigElementList<ClassloaderElement> cel = appRecord.getClassloaders();
+        ClassloaderElement loader = new ClassloaderElement();
+        loader.getCommonLibraryRefs().add("AnonymousJDBCLib");
+        cel.add(loader);
 
         server.setMarkToEndOfLog();
         ServerConfiguration sc = server.getServerConfiguration();
@@ -182,8 +178,7 @@ public class JPA10EmbeddableNested_EJB extends JPAFATServletClient {
             }
 
             server.stopServer("CWWJP9991W", // From Eclipselink drop-and-create tables option
-                              "WTRN0074E: Exception caught from before_completion synchronization operation", // RuntimeException test, expected
-                              "CWWJP0055E" // TODO: OpenJPA throws a MetaDataException while parsing the XML nested embeddables
+                              "WTRN0074E: Exception caught from before_completion synchronization operation" // RuntimeException test, expected
             );
         } finally {
             try {
@@ -197,7 +192,7 @@ public class JPA10EmbeddableNested_EJB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(JPA10EmbeddableNested_EJB.class, timestart);
+            bannerEnd(JPA10EmbeddableBasic_WEB.class, timestart);
         }
     }
 }
