+++ dev/com.ibm.ws.jpa.tests.spec10.embeddable_fat/fat/src/com/ibm/ws/jpa/spec10/embeddable/tests/JPA10EmbeddableNested_EJB.java
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.spec10.embeddable;
+package com.ibm.ws.jpa.spec10.embeddable.tests;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,7 +27,10 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.embeddable.relationship.web.TestEmbeddableRelationshipServlet;
+import com.ibm.ws.jpa.embeddable.nested.ejb.TestEmbeddableNested_EJB_SFEx_Servlet;
+import com.ibm.ws.jpa.embeddable.nested.ejb.TestEmbeddableNested_EJB_SF_Servlet;
+import com.ibm.ws.jpa.embeddable.nested.ejb.TestEmbeddableNested_EJB_SL_Servlet;
+import com.ibm.ws.jpa.spec10.embeddable.FATSuite;
 
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
@@ -42,11 +45,11 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class JPA10EmbeddableRelationship_WEB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "embeddableRelationshipWeb";
-    private final static String RESOURCE_ROOT = "test-applications/embeddable/relationship/";
-    private final static String appFolder = "web";
-    private final static String appName = "embeddableRelationshipWeb";
+public class JPA10EmbeddableNested_EJB extends JPAFATServletClient {
+    private final static String CONTEXT_ROOT = "embeddableNestedEjb";
+    private final static String RESOURCE_ROOT = "test-applications/embeddable/nested/";
+    private final static String appFolder = "ejb";
+    private final static String appName = "embeddableNestedEjb";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
@@ -55,13 +58,15 @@ public class JPA10EmbeddableRelationship_WEB extends JPAFATServletClient {
     private static long timestart = 0;
 
     static {
-        dropSet.add("JPA10_EMBEDDABLE_RELATIONSHIP_DROP_${dbvendor}.ddl");
-        createSet.add("JPA10_EMBEDDABLE_RELATIONSHIP_CREATE_${dbvendor}.ddl");
+        dropSet.add("JPA10_EMBEDDABLE_NESTED_DROP_${dbvendor}.ddl");
+        createSet.add("JPA10_EMBEDDABLE_NESTED_CREATE_${dbvendor}.ddl");
     }
 
     @Server("JPA10Server")
     @TestServlets({
-                    @TestServlet(servlet = TestEmbeddableRelationshipServlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableRelationshipServlet")
+                    @TestServlet(servlet = TestEmbeddableNested_EJB_SL_Servlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableNested_EJB_SL_Servlet"),
+                    @TestServlet(servlet = TestEmbeddableNested_EJB_SF_Servlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableNested_EJB_SF_Servlet"),
+                    @TestServlet(servlet = TestEmbeddableNested_EJB_SFEx_Servlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableNested_EJB_SFEx_Servlet")
     })
     public static LibertyServer server;
 
@@ -70,7 +75,7 @@ public class JPA10EmbeddableRelationship_WEB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(JPA10EmbeddableRelationship_WEB.class);
+        bannerStart(JPA10EmbeddableNested_EJB.class);
         timestart = System.currentTimeMillis();
 
         int appStartTimeout = server.getAppStartTimeout();
@@ -95,7 +100,7 @@ public class JPA10EmbeddableRelationship_WEB extends JPAFATServletClient {
 
         final Set<String> ddlSet = new HashSet<String>();
 
-        System.out.println(JPA10EmbeddableRelationship_WEB.class.getName() + " Setting up database tables...");
+        System.out.println(JPA10EmbeddableNested_EJB.class.getName() + " Setting up database tables...");
 
         ddlSet.clear();
         for (String ddlName : dropSet) {
@@ -113,15 +118,20 @@ public class JPA10EmbeddableRelationship_WEB extends JPAFATServletClient {
     }
 
     private static void setupTestApplication() throws Exception {
+        JavaArchive ejbApp = ShrinkWrap.create(JavaArchive.class, appName + ".jar");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.embeddable.nested.ejblocal");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.embeddable.nested.model");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.embeddable.nested.testlogic");
+        ShrinkHelper.addDirectory(ejbApp, RESOURCE_ROOT + appFolder + "/" + appName + ".jar");
+
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.relationship.model");
-        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.relationship.testlogic");
-        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.relationship.web");
+        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.nested.ejb");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
 
         final EnterpriseArchive app = ShrinkWrap.create(EnterpriseArchive.class, appNameEar);
+        app.addAsModule(ejbApp);
         app.addAsModule(webApp);
         app.addAsLibrary(testApiJar);
         ShrinkHelper.addDirectory(app, RESOURCE_ROOT + appFolder, new org.jboss.shrinkwrap.api.Filter<ArchivePath>() {
@@ -188,7 +198,7 @@ public class JPA10EmbeddableRelationship_WEB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(JPA10EmbeddableRelationship_WEB.class, timestart);
+            bannerEnd(JPA10EmbeddableNested_EJB.class, timestart);
         }
     }
 }
