+++ dev/com.ibm.ws.jpa.tests.spec10.embeddable_fat/fat/src/com/ibm/ws/jpa/spec10/embeddable/tests/JPA10EmbeddableRelationship_WEB.java
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.spec10.embeddable;
+package com.ibm.ws.jpa.spec10.embeddable.tests;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,9 +27,8 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.embeddable.relationship.ejb.TestEmbeddableRelationship_EJB_SFEx_Servlet;
-import com.ibm.ws.jpa.embeddable.relationship.ejb.TestEmbeddableRelationship_EJB_SF_Servlet;
-import com.ibm.ws.jpa.embeddable.relationship.ejb.TestEmbeddableRelationship_EJB_SL_Servlet;
+import com.ibm.ws.jpa.embeddable.relationship.web.TestEmbeddableRelationshipServlet;
+import com.ibm.ws.jpa.spec10.embeddable.FATSuite;
 
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
@@ -44,11 +43,11 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class JPA10EmbeddableRelationship_EJB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "embeddableRelationshipEjb";
+public class JPA10EmbeddableRelationship_WEB extends JPAFATServletClient {
+    private final static String CONTEXT_ROOT = "embeddableRelationshipWeb";
     private final static String RESOURCE_ROOT = "test-applications/embeddable/relationship/";
-    private final static String appFolder = "ejb";
-    private final static String appName = "embeddableRelationshipEjb";
+    private final static String appFolder = "web";
+    private final static String appName = "embeddableRelationshipWeb";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
@@ -63,9 +62,7 @@ public class JPA10EmbeddableRelationship_EJB extends JPAFATServletClient {
 
     @Server("JPA10Server")
     @TestServlets({
-                    @TestServlet(servlet = TestEmbeddableRelationship_EJB_SL_Servlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableRelationship_EJB_SL_Servlet"),
-                    @TestServlet(servlet = TestEmbeddableRelationship_EJB_SF_Servlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableRelationship_EJB_SF_Servlet"),
-                    @TestServlet(servlet = TestEmbeddableRelationship_EJB_SFEx_Servlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableRelationship_EJB_SFEx_Servlet")
+                    @TestServlet(servlet = TestEmbeddableRelationshipServlet.class, path = CONTEXT_ROOT + "/" + "TestEmbeddableRelationshipServlet")
     })
     public static LibertyServer server;
 
@@ -74,7 +71,7 @@ public class JPA10EmbeddableRelationship_EJB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(JPA10EmbeddableRelationship_EJB.class);
+        bannerStart(JPA10EmbeddableRelationship_WEB.class);
         timestart = System.currentTimeMillis();
 
         int appStartTimeout = server.getAppStartTimeout();
@@ -99,7 +96,7 @@ public class JPA10EmbeddableRelationship_EJB extends JPAFATServletClient {
 
         final Set<String> ddlSet = new HashSet<String>();
 
-        System.out.println(JPA10EmbeddableRelationship_EJB.class.getName() + " Setting up database tables...");
+        System.out.println(JPA10EmbeddableRelationship_WEB.class.getName() + " Setting up database tables...");
 
         ddlSet.clear();
         for (String ddlName : dropSet) {
@@ -117,20 +114,15 @@ public class JPA10EmbeddableRelationship_EJB extends JPAFATServletClient {
     }
 
     private static void setupTestApplication() throws Exception {
-        JavaArchive ejbApp = ShrinkWrap.create(JavaArchive.class, appName + ".jar");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.embeddable.relationship.ejblocal");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.embeddable.relationship.model");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.embeddable.relationship.testlogic");
-        ShrinkHelper.addDirectory(ejbApp, RESOURCE_ROOT + appFolder + "/" + appName + ".jar");
-
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.relationship.ejb");
+        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.relationship.model");
+        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.relationship.testlogic");
+        webApp.addPackages(true, "com.ibm.ws.jpa.embeddable.relationship.web");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
 
         final EnterpriseArchive app = ShrinkWrap.create(EnterpriseArchive.class, appNameEar);
-        app.addAsModule(ejbApp);
         app.addAsModule(webApp);
         app.addAsLibrary(testApiJar);
         ShrinkHelper.addDirectory(app, RESOURCE_ROOT + appFolder, new org.jboss.shrinkwrap.api.Filter<ArchivePath>() {
@@ -197,7 +189,7 @@ public class JPA10EmbeddableRelationship_EJB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(JPA10EmbeddableRelationship_EJB.class, timestart);
+            bannerEnd(JPA10EmbeddableRelationship_WEB.class, timestart);
         }
     }
 }
